estate-investment/pyproject.toml
[project]
name = "estate-investment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "folium>=0.20.0",
    "geopy>=2.4.1",
    "pillow>=12.2.0",
    "pytesseract>=0.3.13",
    "requests>=2.33.1",
]
